[multiplatform] multiple file refactor to enable freebsd, solaris, netbsd builds.
