+++ pkg/metadata/host/host_nix.go
@@ -1,0 +1,11 @@
+// +build freebsd netbsd openbsd solaris dragonfly
+
+package host
+
+import "github.com/shirou/gopsutil/host"
+
+type osVersion [3]string
+
+func fillOsVersion(stats *systemStats, info *host.InfoStat) {
+	stats.Nixver = osVersion{info.PlatformFamily, info.PlatformVersion, ""}
+}
