+++ pkg/metadata/v5/v5_other.go
@@ -1,10 +1,9 @@
-// +build linux windows
+// +build freebsd netbsd openbsd solaris dragonfly
 
 package v5
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/metadata/common"
-	"github.com/DataDog/datadog-agent/pkg/metadata/gohai"
 	"github.com/DataDog/datadog-agent/pkg/metadata/host"
 	"github.com/DataDog/datadog-agent/pkg/metadata/resources"
 )
@@ -14,11 +13,9 @@ func GetPayload(hostname string) *Payload {
 	cp := common.GetPayload()
 	hp := host.GetPayload(hostname)
 	rp := resources.GetPayload(hostname)
-	gp := gohai.GetPayload()
 	return &Payload{
 		CommonPayload:    CommonPayload{*cp},
 		HostPayload:      HostPayload{*hp},
 		ResourcesPayload: ResourcesPayload{*rp},
-		GohaiPayload:     GohaiPayload{*gp},
 	}
 }
