+++ pkg/util/hostname_docker_stub.go
@@ -1,0 +1,7 @@
+// +build freebsd netbsd openbsd solaris dragonfly
+
+package util
+
+func getContainerHostname() (bool, string) {
+	return false, ""
+}
